Add linter checks and fix linter warnings (#28)
* Format imports

* Fix nlreturn linter warnings

* Fix godot linter warnings

* Fix unconvert linter warnings

* Fix gosec linter warnings

* Fix nilnil linter warnings

* Fix nilerr linter warnings

* Fix revive linter warnings

* Fix forcetypeassert linter warnings

* Fix ifshort linter warnings

* Fix prealloc linter warnings

* Fix gosimple linter warnings

* Fix errcheck linter warnings

* Fix dupl linter warnings

* Fix gocritic linter warnings

* Fix nosnakecase linter warnings

* Fix scopelint linter warnings

* Add linter settings

* Fix review comments

+++ pkg/participation/event_builder.go
@@ -12,9 +12,9 @@ func NewEventBuilder(name string, milestoneCommence iotago.MilestoneIndex, miles
 	return &EventBuilder{
 		event: &Event{
 			Name:                   name,
-			MilestoneIndexCommence: uint32(milestoneCommence),
-			MilestoneIndexStart:    uint32(milestoneBeginHolding),
-			MilestoneIndexEnd:      uint32(milestoneEnd),
+			MilestoneIndexCommence: milestoneCommence,
+			MilestoneIndexStart:    milestoneBeginHolding,
+			MilestoneIndexEnd:      milestoneEnd,
 			AdditionalInfo:         additionalInfo,
 		},
 	}
@@ -37,9 +37,11 @@ func (rb *EventBuilder) Payload(seri serializer.Serializable) *EventBuilder {
 	case nil:
 	default:
 		rb.err = fmt.Errorf("%w: unsupported type %T", ErrUnknownPayloadType, seri)
+
 		return rb
 	}
 	rb.event.Payload = seri
+
 	return rb
 }
 
@@ -52,5 +54,6 @@ func (rb *EventBuilder) Build() (*Event, error) {
 	if _, err := rb.event.Serialize(serializer.DeSeriModePerformValidation, nil); err != nil {
 		return nil, fmt.Errorf("unable to build participation: %w", err)
 	}
+
 	return rb.event, nil
 }
